feat: update cauldron to storage behaviour
