+++ src/main/java/com/klikli_dev/theurgy/content/apparatus/liquefactioncauldron/LiquefactionCauldronBlockEntity.java
@@ -4,18 +4,13 @@
 
 package com.klikli_dev.theurgy.content.apparatus.liquefactioncauldron;
 
-import com.klikli_dev.theurgy.content.apparatus.distiller.DistillerBlockEntity;
-import com.klikli_dev.theurgy.content.behaviour.CraftingBehaviour;
 import com.klikli_dev.theurgy.content.behaviour.HeatConsumerBehaviour;
-import com.klikli_dev.theurgy.content.behaviour.MonitoredItemStackHandler;
-import com.klikli_dev.theurgy.content.behaviour.PreventInsertWrapper;
 import com.klikli_dev.theurgy.content.capability.DefaultHeatReceiver;
 import com.klikli_dev.theurgy.content.capability.HeatReceiver;
 import com.klikli_dev.theurgy.content.particle.ParticleColor;
 import com.klikli_dev.theurgy.content.particle.coloredbubble.ColoredBubbleParticleProvider;
 import com.klikli_dev.theurgy.registry.BlockEntityRegistry;
 import com.klikli_dev.theurgy.registry.CapabilityRegistry;
-import com.klikli_dev.theurgy.registry.FluidTagRegistry;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
@@ -23,74 +18,36 @@ import net.minecraft.network.Connection;
 import net.minecraft.network.protocol.Packet;
 import net.minecraft.network.protocol.game.ClientGamePacketListener;
 import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.common.capabilities.Capability;
-import net.minecraftforge.common.capabilities.ForgeCapabilities;
 import net.minecraftforge.common.util.LazyOptional;
-import net.minecraftforge.fluids.FluidStack;
-import net.minecraftforge.fluids.FluidType;
-import net.minecraftforge.fluids.capability.IFluidHandler;
-import net.minecraftforge.fluids.capability.templates.FluidTank;
-import net.minecraftforge.items.IItemHandler;
-import net.minecraftforge.items.ItemStackHandler;
-import net.minecraftforge.items.wrapper.CombinedInvWrapper;
-import net.minecraftforge.registries.ForgeRegistries;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.function.Predicate;
-
 public class LiquefactionCauldronBlockEntity extends BlockEntity {
-
-    public ItemStackHandler inputInventory;
-    /**
-     * The underlying outputInventory which allows inserting too - we use this when crafting.
-     */
-    public ItemStackHandler outputInventory;
-    /**
-     * A wrapper that only allows taking from the outputInventory - this is what we show to the outside.
-     */
-    public PreventInsertWrapper outputInventoryTakeOnlyWrapper;
-
-    public CombinedInvWrapper inventory;
-    public LazyOptional<IItemHandler> inventoryCapability;
-    public LazyOptional<IItemHandler> inputInventoryCapability;
-    public LazyOptional<IItemHandler> outputInventoryCapability;
-
-    public FluidTank solventTank;
-    public LazyOptional<IFluidHandler> solventTankCapability;
 
     public DefaultHeatReceiver heatReceiver;
     public LazyOptional<HeatReceiver> heatReceiverCapability;
 
+    public LiquefactionCauldronStorageBehaviour storageBehaviour;
 
-    protected CraftingBehaviour<?, ?, ?> craftingBehaviour;
+
+    protected LiquefactionCauldronCraftingBehaviour craftingBehaviour;
     protected HeatConsumerBehaviour heatConsumerBehaviour;
 
 
     public LiquefactionCauldronBlockEntity(BlockPos pPos, BlockState pBlockState) {
         super(BlockEntityRegistry.LIQUEFACTION_CAULDRON.get(), pPos, pBlockState);
 
-        this.inputInventory = new InputInventory();
-        this.outputInventory = new OutputInventory();
-        this.outputInventoryTakeOnlyWrapper = new PreventInsertWrapper(this.outputInventory);
-        this.inventory = new CombinedInvWrapper(this.inputInventory, this.outputInventoryTakeOnlyWrapper);
 
-        this.inventoryCapability = LazyOptional.of(() -> this.inventory);
-        this.inputInventoryCapability = LazyOptional.of(() -> this.inputInventory);
-        this.outputInventoryCapability = LazyOptional.of(() -> this.outputInventoryTakeOnlyWrapper);
-
-        this.solventTank = new SolventTank(FluidType.BUCKET_VOLUME * 2, (fluidStack -> ForgeRegistries.FLUIDS.tags().getTag(FluidTagRegistry.SOLVENT).contains(fluidStack.getFluid())));
-
-        this.solventTankCapability = LazyOptional.of(() -> this.solventTank);
+        this.storageBehaviour = new LiquefactionCauldronStorageBehaviour(this, () -> this.craftingBehaviour);
 
         this.heatReceiver = new DefaultHeatReceiver();
         this.heatReceiverCapability = LazyOptional.of(() -> this.heatReceiver);
 
-        this.craftingBehaviour = new LiquefactionCauldronCraftingBehaviour(this, () -> this.inputInventory, () -> this.outputInventory, () -> this.solventTank);
+        this.craftingBehaviour = new LiquefactionCauldronCraftingBehaviour(this, () -> this.storageBehaviour.inputInventory, () -> this.storageBehaviour.outputInventory, () -> this.storageBehaviour.solventTank);
         this.heatConsumerBehaviour = new HeatConsumerBehaviour(this);
     }
 
@@ -121,16 +78,12 @@ public class LiquefactionCauldronBlockEntity extends BlockEntity {
     }
 
     public void readNetwork(CompoundTag tag) {
-        if (tag.contains("solventTank")) {
-            this.solventTank.readFromNBT(tag.getCompound("solventTank"));
-        }
-
+        this.storageBehaviour.readNetwork(tag);
         this.craftingBehaviour.readNetwork(tag);
     }
 
     public void writeNetwork(CompoundTag tag) {
-        tag.put("solventTank", this.solventTank.writeToNBT(new CompoundTag()));
-
+        this.storageBehaviour.writeNetwork(tag);
         this.craftingBehaviour.writeNetwork(tag);
     }
 
@@ -141,7 +94,7 @@ public class LiquefactionCauldronBlockEntity extends BlockEntity {
 
     public void tickServer() {
         boolean isHeated = this.heatConsumerBehaviour.isHeated();
-        boolean hasInput = !this.inputInventory.getStackInSlot(0).isEmpty();
+        boolean hasInput = !this.storageBehaviour.inputInventory.getStackInSlot(0).isEmpty();
 
         this.craftingBehaviour.tickServer(isHeated, hasInput);
     }
@@ -165,13 +118,9 @@ public class LiquefactionCauldronBlockEntity extends BlockEntity {
 
     @Override
     public @NotNull <T> LazyOptional<T> getCapability(@NotNull Capability<T> cap, @Nullable Direction side) {
-        if (cap == ForgeCapabilities.ITEM_HANDLER) {
-            if (side == Direction.UP) return this.inputInventoryCapability.cast();
-            if (side == Direction.DOWN) return this.outputInventoryCapability.cast();
-            return this.inventoryCapability.cast();
-        }
-
-        if (cap == ForgeCapabilities.FLUID_HANDLER) return this.solventTankCapability.cast();
+        var storage = this.storageBehaviour.getCapability(cap, side);
+        if (storage != null)
+            return storage;
 
         if (cap == CapabilityRegistry.HEAT_RECEIVER) {
             return this.heatReceiverCapability.cast();
@@ -183,10 +132,6 @@ public class LiquefactionCauldronBlockEntity extends BlockEntity {
     @Override
     public void invalidateCaps() {
         super.invalidateCaps();
-        this.inventoryCapability.invalidate();
-        this.inputInventoryCapability.invalidate();
-        this.outputInventoryCapability.invalidate();
-        this.solventTankCapability.invalidate();
         this.heatReceiverCapability.invalidate();
     }
 
@@ -194,11 +139,9 @@ public class LiquefactionCauldronBlockEntity extends BlockEntity {
     protected void saveAdditional(CompoundTag pTag) {
         super.saveAdditional(pTag);
 
-        pTag.put("inputInventory", this.inputInventory.serializeNBT());
-        pTag.put("outputInventory", this.outputInventory.serializeNBT());
-        pTag.put("solventTank", this.solventTank.writeToNBT(new CompoundTag()));
         pTag.put("heatReceiver", this.heatReceiver.serializeNBT());
 
+        this.storageBehaviour.saveAdditional(pTag);
         this.craftingBehaviour.saveAdditional(pTag);
     }
 
@@ -206,80 +149,10 @@ public class LiquefactionCauldronBlockEntity extends BlockEntity {
     public void load(CompoundTag pTag) {
         super.load(pTag);
 
-        if (pTag.contains("inputInventory")) this.inputInventory.deserializeNBT(pTag.getCompound("inputInventory"));
-        if (pTag.contains("outputInventory")) this.outputInventory.deserializeNBT(pTag.getCompound("outputInventory"));
-
-        if (pTag.contains("solventTank")) {
-            this.solventTank.readFromNBT(pTag.getCompound("solventTank"));
-        }
-
         if (pTag.contains("heatReceiver"))
             this.heatReceiver.deserializeNBT(pTag.get("heatReceiver"));
 
+        this.storageBehaviour.load(pTag);
         this.craftingBehaviour.load(pTag);
-    }
-
-    public class SolventTank extends FluidTank {
-
-        public SolventTank(int capacity, Predicate<FluidStack> validator) {
-            super(capacity, validator);
-        }
-
-        @Override
-        protected void onContentsChanged() {
-            LiquefactionCauldronBlockEntity.this.setChanged();
-            LiquefactionCauldronBlockEntity.this.sendBlockUpdated();
-        }
-    }
-
-
-    public class InputInventory extends MonitoredItemStackHandler {
-
-        public InputInventory() {
-            super(1);
-        }
-
-        @Override
-        protected void onSetStackInSlot(int slot, ItemStack oldStack, ItemStack newStack, boolean isSameItem) {
-            if(!isSameItem){
-                LiquefactionCauldronBlockEntity.this.craftingBehaviour.onInputItemChanged(oldStack, newStack);
-            }
-        }
-
-        @Override
-        protected void onInsertItem(int slot, ItemStack oldStack, ItemStack newStack, ItemStack toInsert, ItemStack remaining) {
-            if (remaining != newStack) {
-                LiquefactionCauldronBlockEntity.this.craftingBehaviour.onInputItemChanged(oldStack, newStack);
-            }
-        }
-
-        @Override
-        protected void onExtractItem(int slot, ItemStack oldStack, ItemStack newStack, ItemStack extracted) {
-            if(newStack.isEmpty()){
-                LiquefactionCauldronBlockEntity.this.craftingBehaviour.onInputItemChanged(oldStack, newStack);
-            }
-        }
-
-        @Override
-        public boolean isItemValid(int slot, ItemStack stack) {
-            return LiquefactionCauldronBlockEntity.this.craftingBehaviour.canProcess(stack) && super.isItemValid(slot, stack);
-        }
-
-        @Override
-        protected void onContentsChanged(int slot) {
-            LiquefactionCauldronBlockEntity.this.setChanged();
-        }
-    }
-
-    public class OutputInventory extends ItemStackHandler {
-
-        public OutputInventory() {
-            super(1);
-        }
-
-        @Override
-        protected void onContentsChanged(int slot) {
-            LiquefactionCauldronBlockEntity.this.setChanged();
-        }
     }
 }
