+++ src/main/java/com/klikli_dev/theurgy/content/apparatus/liquefactioncauldron/LiquefactionCauldronStorageBehaviour.java
@@ -1,0 +1,159 @@
+package com.klikli_dev.theurgy.content.apparatus.liquefactioncauldron;
+
+import com.klikli_dev.theurgy.content.behaviour.MonitoredItemStackHandler;
+import com.klikli_dev.theurgy.content.behaviour.PreventInsertWrapper;
+import com.klikli_dev.theurgy.content.behaviour.StorageBehaviour;
+import com.klikli_dev.theurgy.registry.FluidTagRegistry;
+import net.minecraft.core.Direction;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.ForgeCapabilities;
+import net.minecraftforge.common.util.LazyOptional;
+import net.minecraftforge.fluids.FluidStack;
+import net.minecraftforge.fluids.FluidType;
+import net.minecraftforge.fluids.capability.IFluidHandler;
+import net.minecraftforge.fluids.capability.templates.FluidTank;
+import net.minecraftforge.items.IItemHandler;
+import net.minecraftforge.items.ItemStackHandler;
+import net.minecraftforge.items.wrapper.CombinedInvWrapper;
+import net.minecraftforge.registries.ForgeRegistries;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.function.Predicate;
+import java.util.function.Supplier;
+
+public class LiquefactionCauldronStorageBehaviour extends StorageBehaviour<LiquefactionCauldronStorageBehaviour> {
+
+    public ItemStackHandler inputInventory;
+    /**
+     * The underlying outputInventory which allows inserting too - we use this when crafting.
+     */
+    public ItemStackHandler outputInventory;
+    /**
+     * A wrapper that only allows taking from the outputInventory - this is what we show to the outside.
+     */
+    public PreventInsertWrapper outputInventoryTakeOnlyWrapper;
+
+    public CombinedInvWrapper inventory;
+    public LazyOptional<IItemHandler> inventoryCapability;
+    public LazyOptional<IItemHandler> inputInventoryCapability;
+    public LazyOptional<IItemHandler> outputInventoryCapability;
+
+    public FluidTank solventTank;
+    public LazyOptional<IFluidHandler> solventTankCapability;
+
+    public Supplier<LiquefactionCauldronCraftingBehaviour> craftingBehaviour;
+
+    public LiquefactionCauldronStorageBehaviour(BlockEntity blockEntity, Supplier<LiquefactionCauldronCraftingBehaviour> craftingBehaviour) {
+        super(blockEntity);
+
+        this.craftingBehaviour = craftingBehaviour;
+
+        this.inputInventory = new InputInventory();
+        this.outputInventory = new OutputInventory();
+        this.outputInventoryTakeOnlyWrapper = new PreventInsertWrapper(this.outputInventory);
+        this.inventory = new CombinedInvWrapper(this.inputInventory, this.outputInventoryTakeOnlyWrapper);
+
+        this.inventoryCapability = LazyOptional.of(() -> this.inventory);
+        this.inputInventoryCapability = LazyOptional.of(() -> this.inputInventory);
+        this.outputInventoryCapability = LazyOptional.of(() -> this.outputInventoryTakeOnlyWrapper);
+
+        this.solventTank = new SolventTank(FluidType.BUCKET_VOLUME * 2, (fluidStack -> ForgeRegistries.FLUIDS.tags().getTag(FluidTagRegistry.SOLVENT).contains(fluidStack.getFluid())));
+
+        this.solventTankCapability = LazyOptional.of(() -> this.solventTank);
+    }
+
+    @Override
+    public void readNetwork(CompoundTag pTag) {
+        if (pTag.contains("inputInventory")) this.inputInventory.deserializeNBT(pTag.getCompound("inputInventory"));
+        if (pTag.contains("outputInventory")) this.outputInventory.deserializeNBT(pTag.getCompound("outputInventory"));
+        if (pTag.contains("solventTank")) this.solventTank.readFromNBT(pTag.getCompound("solventTank"));
+    }
+
+    @Override
+    public void writeNetwork(CompoundTag pTag) {
+        pTag.put("inputInventory", this.inputInventory.serializeNBT());
+        pTag.put("outputInventory", this.outputInventory.serializeNBT());
+        pTag.put("solventTank", this.solventTank.writeToNBT(new CompoundTag()));
+    }
+
+    @Override
+    public void saveAdditional(CompoundTag pTag) {
+        pTag.put("inputInventory", this.inputInventory.serializeNBT());
+        pTag.put("outputInventory", this.outputInventory.serializeNBT());
+        pTag.put("solventTank", this.solventTank.writeToNBT(new CompoundTag()));
+    }
+
+    @Override
+    public void load(CompoundTag pTag) {
+        if (pTag.contains("inputInventory")) this.inputInventory.deserializeNBT(pTag.getCompound("inputInventory"));
+        if (pTag.contains("outputInventory")) this.outputInventory.deserializeNBT(pTag.getCompound("outputInventory"));
+        if (pTag.contains("solventTank")) this.solventTank.readFromNBT(pTag.getCompound("solventTank"));
+    }
+
+    @Override
+    public <T> LazyOptional<T> getCapability(@NotNull Capability<T> cap, @Nullable Direction side) {
+        if (cap == ForgeCapabilities.ITEM_HANDLER) {
+            if (side == Direction.UP) return this.inputInventoryCapability.cast();
+            if (side == Direction.DOWN) return this.outputInventoryCapability.cast();
+            return this.inventoryCapability.cast();
+        }
+
+        if (cap == ForgeCapabilities.FLUID_HANDLER) return this.solventTankCapability.cast();
+
+        return null;
+    }
+
+    public class SolventTank extends FluidTank {
+
+        public SolventTank(int capacity, Predicate<FluidStack> validator) {
+            super(capacity, validator);
+        }
+
+        @Override
+        protected void onContentsChanged() {
+            LiquefactionCauldronStorageBehaviour.this.setChanged();
+            LiquefactionCauldronStorageBehaviour.this.sendBlockUpdated();
+        }
+    }
+
+    public class InputInventory extends MonitoredItemStackHandler {
+
+        public InputInventory() {
+            super(1);
+        }
+
+        @Override
+        protected void onContentTypeChanged(int slot, ItemStack oldStack, ItemStack newStack) {
+            LiquefactionCauldronStorageBehaviour.this.craftingBehaviour.get().onInputItemChanged(oldStack, newStack);
+            //we also need to network sync our BE, because if the content type changes then the interaction behaviour client side changes
+            LiquefactionCauldronStorageBehaviour.this.sendBlockUpdated();
+        }
+
+
+        @Override
+        public boolean isItemValid(int slot, ItemStack stack) {
+            return LiquefactionCauldronStorageBehaviour.this.craftingBehaviour.get().canProcess(stack) && super.isItemValid(slot, stack);
+        }
+
+        @Override
+        protected void onContentsChanged(int slot) {
+            LiquefactionCauldronStorageBehaviour.this.setChanged();
+        }
+    }
+
+    public class OutputInventory extends ItemStackHandler {
+
+        public OutputInventory() {
+            super(1);
+        }
+
+        @Override
+        protected void onContentsChanged(int slot) {
+            LiquefactionCauldronStorageBehaviour.this.setChanged();
+        }
+    }
+}
